Removed cross versioning from snap. We don't need that krap!
Actually, it was causing issues, so removing it fixes an issue with finding the UI.  Also,
we don't really need cross versioning since we won't be cross-releasing against various
scala versions.   We're like SBT here.   Light up a cigar and laugh when someone mentions
binary compatiblilty and Scala.  We're well beyond the point of no return.