Initial cut at controller fed plugins for the UI.
This brings with it the basic notion of having a Play controller generate the require.js javascript
of which plugins to load with the application, requiring a real refresh if we every add a new plugin.
We still need to work out actual plugin hosting and some kind of plugin manager that tells us which
'entry point' to load for a given plugin, but we're approaching zen.