Added a basic 'view' functionality to the code plugin that allows us to view the contents of a file.   Right now, there are two types of files supported, 'code' and 'image', everything else goes to a 'cannot display' screen.
While this *works*, it leaves a lot to be desired.  CSS styling, for one.  Also,
there's a really wonky interaction between the file BROWSING and the VIEWING.  Mostly,
because I'm not sure how to design the URL schema such that I *know* that we can have the
browse widgets *and* a view widget beside them.    Until we figure out all this
annoying details, this is what we'll have for now.

Review / Fix up look by @Warry