Initial newApp form now implemented and wired w/ play's form stuff.
This does not handle failures or flash messages or any niceties.  Probably
someone needs to teach me that stuff, or I need to read some more.  For now,
I'm going to finish wiring in the actual creation of a new project in the directory
specified, and all then push to the @havocp side of the fence the registering of
a project in the app history cache and such.