Fixes MACOSX double-click UI launch.
While we can't detect if a script was double clicked in ACTUALITY, we have an advantage:

The PWD is set to the home directory.

SO, this commit assumes that if you're launching from the home directory, you *Really* want
the UI.   Another thing that appears to be true, but I didn't code in the logic, is that
the path of the $0 is *complete*.   If we need to add that restirction as well, we can.

Review by @jamesward