A very hacky first cut at loading applications from locations.
This uses the file browser on the home page to browse to a project directory and attempt to load that directory into builder.   We add configurability to the file selection widget so we can specify the button names.   We also tweak some CSS so things are visible.   Finally, we add a new route where you can pass in a location and get back the 'id' of the application so you can load that builder URL.