Introduce sbt boot events and transparently handle reboot
Split sbt child actor into a supervisor wrapper and an underlying,
so that the underlying can be restarted if needed.

Only restart the underlying if it failed due to a deliberate
request to reboot from our remote probe.

Only send protocol.Started once we've successfully booted
up, potentially on the second try.