Update shim config in the parent process
This required factoring out an "io" module which
is compiled with both Scala versions and used by
both remote probe and the "common" module.

The shim checking is set up so there's just one
canonical list of shim names used by both the
parent and the probe, to avoid confusion.

The parent does the update in child pool, before
spawning any sbts, so we don't have to check
every time we launch a new sbt instance.