Clump process output within 30ms into a single event
This way if Play or whatever writes out a line in multiple
system calls, we will tend to clump it back up into a single
line for processing. Nothing breaks if the clumping fails,
but it's prettier when it succeeds, which should be almost
always.

We could just wait for newlines to flush but there are
ways that could break too (like never getting a newline
or not displaying "Loading... done" type stuff).