Add UsersActorSpec, add some documentation
Added specs for UsersActor, covering how to inject TestProbe into a
child actor using creator.  Had to refactor the UsersActor a bit to
allow for the injection, and  had to break out the support classes from
UserActorSpec.

Also added some javadoc to the classes explaining how to turn on
debugging in Akka and why there are both Scala and Akka actors.