Fix UserConfig to reload lazily
When we got a load failure, we were immediately kicking off
an async reload. The tests depended on this async reload completing
before a file.delete(). Anyway the immediate kickoff behavior seems
dumb, better to retry loading the file "later" (next time someone
wants it). Fixed the tests to check the new behavior.