Fix browser reload behavior
The problem was that when you reload, the closed websocket kills the
app actor, but not before the new reloaded page has tried to use the
old app actor.

So now when loading /app we poll the app actor for a few seconds
until we get one that says it has never had a connected socket.

Somewhat unrelated, some of the '?' calls have a DeathReportingProxy
inserted in the middle, which prevents some confusing long timeouts.
In this new setup, the long timeouts hypothetically won't happen
anyway, though. But nice to fix regardless.